fix(mimo-v2): use float32 for MoE router gate to match SGLang precision
SGLang's MoEGate hardcodes float32 for gate weight and casts
hidden_states to float32 before the gate matmul. sgl-jax was using
bfloat16, causing precision loss in routing scores that can flip
expert selection for tokens with close scores.

+++ python/sgl_jax/srt/models/mimo_v2_flash.py
@@ -89,7 +89,7 @@ class MiMoV2Moe(nnx.Module):
         self.moe_gate = GateLogit(
             input_size=config.hidden_size,
             num_experts=num_experts,
-            weight_dtype=dtype,
+            weight_dtype=jnp.float32,
             score_func=getattr(config, "scoring_func", "softmax"),
         )
